Create FrontiersSummarizer_Testing_v1.ipynb
Major update to scraping code as different HTML formats were identified, seemingly distinguished by a certain year (approximately between 2008-2012) of the paper.  Different scraping method used depending on paper HTML format.

Combined everything into classes:
1) Connect: creates connection to the link
2) Scrape: scrapes the information from the link
3) Clean_Filter: converts html format to just text

Finally, a testing function was added to test the scraping and cleaning methods mainly on a large list of links, verifying the robustness of the program.

Note: more methods will be added as the complete list annotated in the notebook is not all in the code.

Additionally, more refined regex needs to be in place to remove white spaces between paragraphs (i.e. '\n' list items in dictionary).